requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
bitarray==2.8.3
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
clarabel==0.6.0
click==8.1.7
cvxpy==1.4.1
cytoolz==0.12.2
dataclasses-json==0.6.2
distro==1.8.0
ecos==2.0.12
eth-abi==4.2.1
eth-account==0.10.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.5.2
eth-utils==2.3.1
exceptiongroup==1.1.3
flask==3.0.0
Flask-Cors==4.0.0
frozenlist==1.4.0
greenlet==3.0.1
h11==0.14.0
hexbytes==0.3.1
httpcore==1.0.2
httpx==0.25.1
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
langchain==0.0.337
langsmith==0.0.64
lru-dict==1.2.0
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
openai==1.3.2
osqp==0.6.3
packaging==23.2
pandas==2.1.3
parsimonious==0.9.0
patsy==0.5.3
protobuf==4.25.1
pybind11==2.11.1
pycryptodome==3.19.0
pydantic==2.5.1
pydantic-core==2.14.3
pyportfolioopt==1.5.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyunormalize==15.1.0
PyYAML==6.0.1
qdldl==0.1.7.post0
referencing==0.31.0
regex==2023.10.3
requests==2.31.0
rlp==3.0.0
rpds-py==0.13.0
scikit-learn==1.3.2
scipy==1.11.3
scs==3.2.4
six==1.16.0
sklearn==0.0.post11
sniffio==1.3.0
SQLAlchemy==2.0.23
statsmodels==0.14.0
tenacity==8.2.3
threadpoolctl==3.2.0
toolz==0.12.0
tqdm==4.66.1
typing-extensions==4.8.0
typing-inspect==0.9.0
tzdata==2023.3
urllib3==2.1.0
web3==6.11.3
websockets==12.0
werkzeug==3.0.1
yarl==1.9.2
zipp==3.17.0
